Time: 104 ms (93.76%), Space: 40.2 MB (84.92%) - LeetHub

+++ 0907-sum-of-subarray-minimums/0907-sum-of-subarray-minimums.cpp
@@ -34,7 +34,7 @@ public:
         
         int maxA = 0;
         for (int i = 0; i < n; i++) {
-            long long prod = (lt[i]*rt[i])%mod;
+            long prod = (lt[i]*rt[i])%mod;
             maxA = (maxA + (prod*arr[i])%mod)%mod;
         }
         return maxA;
